refactor(discord): rename the class to have a match between the file name and the class name
