+++ src/discord/discord-author.ts
